Refactor trace writer trace dir helper
