+++ crates/rulemorph_trace/src/trace_writer/trace_dir.rs
@@ -1,0 +1,49 @@
+use std::fs;
+use std::path::{Path, PathBuf};
+use std::time::{SystemTime, UNIX_EPOCH};
+
+use anyhow::Result;
+use tracing::warn;
+
+pub(super) fn ensure_unique_trace_dir(
+    base_dir: &Path,
+    trace_id: String,
+    raw_trace_id: &str,
+) -> Result<(String, PathBuf)> {
+    fs::create_dir_all(base_dir)?;
+    let base_id = trace_id;
+    let nanos = SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap_or_default()
+        .as_nanos();
+    let mut counter = 0usize;
+    loop {
+        let candidate = if counter == 0 {
+            base_id.clone()
+        } else {
+            let suffix = if counter == 1 {
+                format!("dup-{nanos}")
+            } else {
+                format!("dup-{nanos}-{}", counter - 1)
+            };
+            format!("{base_id}-{suffix}")
+        };
+        let trace_dir = base_dir.join(&candidate);
+        match fs::create_dir(&trace_dir) {
+            Ok(()) => {
+                if counter > 0 {
+                    warn!(
+                        "trace_id collision for {}; using {} instead",
+                        raw_trace_id, candidate
+                    );
+                }
+                return Ok((candidate, trace_dir));
+            }
+            Err(err) if err.kind() == std::io::ErrorKind::AlreadyExists => {
+                counter = counter.saturating_add(1);
+                continue;
+            }
+            Err(err) => return Err(err.into()),
+        }
+    }
+}
